+++ src/Symfony/Component/HttpFoundation/Tests/ParameterBagTest.php
@@ -12,12 +12,15 @@
 namespace Symfony\Component\HttpFoundation\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 use Symfony\Component\HttpFoundation\Exception\BadRequestException;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Tests\Fixtures\FooEnum;
 
 class ParameterBagTest extends TestCase
 {
+    use ExpectDeprecationTrait;
+
     public function testConstructor()
     {
         $this->testAll();
@@ -112,34 +115,137 @@ class ParameterBagTest extends TestCase
 
     public function testGetAlpha()
     {
-        $bag = new ParameterBag(['word' => 'foo_BAR_012']);
+        $bag = new ParameterBag(['word' => 'foo_BAR_012', 'bool' => true, 'integer' => 123]);
 
-        $this->assertEquals('fooBAR', $bag->getAlpha('word'), '->getAlpha() gets only alphabetic characters');
-        $this->assertEquals('', $bag->getAlpha('unknown'), '->getAlpha() returns empty string if a parameter is not defined');
+        $this->assertSame('fooBAR', $bag->getAlpha('word'), '->getAlpha() gets only alphabetic characters');
+        $this->assertSame('', $bag->getAlpha('unknown'), '->getAlpha() returns empty string if a parameter is not defined');
+        $this->assertSame('abcDEF', $bag->getAlpha('unknown', 'abc_DEF_012'), '->getAlpha() returns filtered default if a parameter is not defined');
+        $this->assertSame('', $bag->getAlpha('integer', 'abc_DEF_012'), '->getAlpha() returns empty string if a parameter is an integer');
+        $this->assertSame('', $bag->getAlpha('bool', 'abc_DEF_012'), '->getAlpha() returns empty string if a parameter is a boolean');
+    }
+
+    public function testGetAlphaExceptionWithArray()
+    {
+        $bag = new ParameterBag(['word' => ['foo_BAR_012']]);
+
+        $this->expectException(\UnexpectedValueException::class);
+        $this->expectExceptionMessage('Parameter value "word" cannot be converted to "string".');
+
+        $bag->getAlpha('word');
     }
 
     public function testGetAlnum()
     {
-        $bag = new ParameterBag(['word' => 'foo_BAR_012']);
+        $bag = new ParameterBag(['word' => 'foo_BAR_012', 'bool' => true, 'integer' => 123]);
 
-        $this->assertEquals('fooBAR012', $bag->getAlnum('word'), '->getAlnum() gets only alphanumeric characters');
-        $this->assertEquals('', $bag->getAlnum('unknown'), '->getAlnum() returns empty string if a parameter is not defined');
+        $this->assertSame('fooBAR012', $bag->getAlnum('word'), '->getAlnum() gets only alphanumeric characters');
+        $this->assertSame('', $bag->getAlnum('unknown'), '->getAlnum() returns empty string if a parameter is not defined');
+        $this->assertSame('abcDEF012', $bag->getAlnum('unknown', 'abc_DEF_012'), '->getAlnum() returns filtered default if a parameter is not defined');
+        $this->assertSame('123', $bag->getAlnum('integer', 'abc_DEF_012'), '->getAlnum() returns the number as string if a parameter is an integer');
+        $this->assertSame('1', $bag->getAlnum('bool', 'abc_DEF_012'), '->getAlnum() returns 1 if a parameter is true');
+    }
+
+    public function testGetAlnumExceptionWithArray()
+    {
+        $bag = new ParameterBag(['word' => ['foo_BAR_012']]);
+
+        $this->expectException(\UnexpectedValueException::class);
+        $this->expectExceptionMessage('Parameter value "word" cannot be converted to "string".');
+
+        $bag->getAlnum('word');
     }
 
     public function testGetDigits()
     {
-        $bag = new ParameterBag(['word' => 'foo_BAR_012']);
+        $bag = new ParameterBag(['word' => 'foo_BAR_0+1-2', 'bool' => true, 'integer' => 123]);
 
-        $this->assertEquals('012', $bag->getDigits('word'), '->getDigits() gets only digits as string');
-        $this->assertEquals('', $bag->getDigits('unknown'), '->getDigits() returns empty string if a parameter is not defined');
+        $this->assertSame('012', $bag->getDigits('word'), '->getDigits() gets only digits as string');
+        $this->assertSame('', $bag->getDigits('unknown'), '->getDigits() returns empty string if a parameter is not defined');
+        $this->assertSame('012', $bag->getDigits('unknown', 'abc_DEF_012'), '->getDigits() returns filtered default if a parameter is not defined');
+        $this->assertSame('123', $bag->getDigits('integer', 'abc_DEF_012'), '->getDigits() returns the number as string if a parameter is an integer');
+        $this->assertSame('1', $bag->getDigits('bool', 'abc_DEF_012'), '->getDigits() returns 1 if a parameter is true');
+    }
+
+    public function testGetDigitsExceptionWithArray()
+    {
+        $bag = new ParameterBag(['word' => ['foo_BAR_012']]);
+
+        $this->expectException(\UnexpectedValueException::class);
+        $this->expectExceptionMessage('Parameter value "word" cannot be converted to "string".');
+
+        $bag->getDigits('word');
     }
 
     public function testGetInt()
     {
-        $bag = new ParameterBag(['digits' => '0123']);
+        $bag = new ParameterBag(['digits' => '123', 'bool' => true]);
 
-        $this->assertEquals(123, $bag->getInt('digits'), '->getInt() gets a value of parameter as integer');
-        $this->assertEquals(0, $bag->getInt('unknown'), '->getInt() returns zero if a parameter is not defined');
+        $this->assertSame(123, $bag->getInt('digits', 0), '->getInt() gets a value of parameter as integer');
+        $this->assertSame(0, $bag->getInt('unknown', 0), '->getInt() returns zero if a parameter is not defined');
+        $this->assertSame(10, $bag->getInt('unknown', 10), '->getInt() returns the default if a parameter is not defined');
+        $this->assertSame(1, $bag->getInt('bool', 0), '->getInt() returns 1 if a parameter is true');
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testGetIntExceptionWithArray()
+    {
+        $this->expectDeprecation('Since symfony/http-foundation 6.3: Ignoring invalid values when using "Symfony\Component\HttpFoundation\ParameterBag::getInt(\'digits\')" is deprecated and will throw an "UnexpectedValueException" in 7.0; use method "filter()" with flag "FILTER_NULL_ON_FAILURE" to keep ignoring them.');
+
+        $bag = new ParameterBag(['digits' => ['123']]);
+        $result = $bag->getInt('digits', 0);
+        $this->assertSame(0, $result);
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testGetIntExceptionWithInvalid()
+    {
+        $this->expectDeprecation('Since symfony/http-foundation 6.3: Ignoring invalid values when using "Symfony\Component\HttpFoundation\ParameterBag::getInt(\'word\')" is deprecated and will throw an "UnexpectedValueException" in 7.0; use method "filter()" with flag "FILTER_NULL_ON_FAILURE" to keep ignoring them.');
+
+        $bag = new ParameterBag(['word' => 'foo_BAR_012']);
+        $result = $bag->getInt('word', 0);
+        $this->assertSame(0, $result);
+    }
+
+    public function testGetString()
+    {
+        $bag = new ParameterBag(['integer' => 123, 'bool_true' => true, 'bool_false' => false, 'string' => 'abc', 'stringable' => new class() implements \Stringable {
+            public function __toString(): string
+            {
+                return 'strval';
+            }
+        }]);
+
+        $this->assertSame('123', $bag->getString('integer'), '->getString() gets a value of parameter as string');
+        $this->assertSame('abc', $bag->getString('string'), '->getString() gets a value of parameter as string');
+        $this->assertSame('', $bag->getString('unknown'), '->getString() returns zero if a parameter is not defined');
+        $this->assertSame('foo', $bag->getString('unknown', 'foo'), '->getString() returns the default if a parameter is not defined');
+        $this->assertSame('1', $bag->getString('bool_true'), '->getString() returns "1" if a parameter is true');
+        $this->assertSame('', $bag->getString('bool_false', 'foo'), '->getString() returns an empty empty string if a parameter is false');
+        $this->assertSame('strval', $bag->getString('stringable'), '->getString() gets a value of a stringable paramater as string');
+    }
+
+    public function testGetStringExceptionWithArray()
+    {
+        $bag = new ParameterBag(['key' => ['abc']]);
+
+        $this->expectException(\UnexpectedValueException::class);
+        $this->expectExceptionMessage('Parameter value "key" cannot be converted to "string".');
+
+        $bag->getString('key');
+    }
+
+    public function testGetStringExceptionWithObject()
+    {
+        $bag = new ParameterBag(['object' => $this]);
+
+        $this->expectException(\UnexpectedValueException::class);
+        $this->expectExceptionMessage('Parameter value "object" cannot be converted to "string".');
+
+        $bag->getString('object');
     }
 
     public function testFilter()
@@ -164,13 +270,13 @@ class ParameterBagTest extends TestCase
         // This test is repeated for code-coverage
         $this->assertEquals('http://example.com/foo', $bag->filter('url', '', \FILTER_VALIDATE_URL, \FILTER_FLAG_PATH_REQUIRED), '->filter() gets a value of parameter as URL with a path');
 
-        $this->assertFalse($bag->filter('dec', '', \FILTER_VALIDATE_INT, [
-            'flags' => \FILTER_FLAG_ALLOW_HEX,
+        $this->assertNull($bag->filter('dec', '', \FILTER_VALIDATE_INT, [
+            'flags' => \FILTER_FLAG_ALLOW_HEX | \FILTER_NULL_ON_FAILURE,
             'options' => ['min_range' => 1, 'max_range' => 0xFF],
         ]), '->filter() gets a value of parameter as integer between boundaries');
 
-        $this->assertFalse($bag->filter('hex', '', \FILTER_VALIDATE_INT, [
-            'flags' => \FILTER_FLAG_ALLOW_HEX,
+        $this->assertNull($bag->filter('hex', '', \FILTER_VALIDATE_INT, [
+            'flags' => \FILTER_FLAG_ALLOW_HEX | \FILTER_NULL_ON_FAILURE,
             'options' => ['min_range' => 1, 'max_range' => 0xFF],
         ]), '->filter() gets a value of parameter as integer between boundaries');
 
@@ -218,12 +324,25 @@ class ParameterBagTest extends TestCase
 
     public function testGetBoolean()
     {
-        $parameters = ['string_true' => 'true', 'string_false' => 'false'];
+        $parameters = ['string_true' => 'true', 'string_false' => 'false', 'string' => 'abc'];
         $bag = new ParameterBag($parameters);
 
         $this->assertTrue($bag->getBoolean('string_true'), '->getBoolean() gets the string true as boolean true');
         $this->assertFalse($bag->getBoolean('string_false'), '->getBoolean() gets the string false as boolean false');
         $this->assertFalse($bag->getBoolean('unknown'), '->getBoolean() returns false if a parameter is not defined');
+        $this->assertTrue($bag->getBoolean('unknown', true), '->getBoolean() returns default if a parameter is not defined');
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testGetBooleanExceptionWithInvalid()
+    {
+        $this->expectDeprecation('Since symfony/http-foundation 6.3: Ignoring invalid values when using "Symfony\Component\HttpFoundation\ParameterBag::getBoolean(\'invalid\')" is deprecated and will throw an "UnexpectedValueException" in 7.0; use method "filter()" with flag "FILTER_NULL_ON_FAILURE" to keep ignoring them.');
+
+        $bag = new ParameterBag(['invalid' => 'foo']);
+        $result = $bag->getBoolean('invalid', 0);
+        $this->assertFalse($result);
     }
 
     public function testGetEnum()
@@ -258,5 +377,17 @@ class ParameterBagTest extends TestCase
         $this->expectExceptionMessage('Parameter "invalid-value" cannot be converted to enum: Symfony\Component\HttpFoundation\Tests\Fixtures\FooEnum::from(): Argument #1 ($value) must be of type int, array given.');
 
         $this->assertNull($bag->getEnum('invalid-value', FooEnum::class));
+    }
+}
+
+class InputStringable
+{
+    public function __construct(private string $value)
+    {
+    }
+
+    public function __toString(): string
+    {
+        return $this->value;
     }
 }